+++ src/buttons.js
@@ -4,10 +4,6 @@
  * proprietary information of Rapid7.
  ******************************************************************************/
 
-import {
-    default as getPrefixer
-} from "./prefixer";
-import utils from "./utils";
 import variables from "./variables";
 
 let buttons = {
@@ -27,7 +23,7 @@ let buttons = {
         textDecoration:"none",
         textTransform:"uppercase",
         transition:"background-color " + variables.transitionTiming + " " + variables.transitionEase + ", color " +
-        variables.transitionTiming + " " + variables.transitionEase,
+            variables.transitionTiming + " " + variables.transitionEase,
         userSelect:"none",
         verticalAlign:"middle",
         whiteSpace:"nowrap",
@@ -40,31 +36,7 @@ let buttons = {
         ":hover":{
             backgroundColor:"#d5d5d5"
         }
-    },
-    buttonDisabled:{
-        border:0,
-        cursor:"not-allowed",
-        opacity:0.4
     }
-};
-
-buttons.disableButton = function(button) {
-    let prefix = getPrefixer();
-
-    return prefix(utils.merge(button, buttons.buttonDisabled, {
-        ":active":{
-            backgroundColor:button.backgroundColor,
-            color:button.color
-        },
-        ":focus":{
-            backgroundColor:button.backgroundColor,
-            color:button.color
-        },
-        ":hover":{
-            backgroundColor:button.backgroundColor,
-            color:button.color
-        }
-    }));
 };
 
 export default buttons;